requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
pytest==7.3.1
pymysql==1.0.3
requests==2.31.0
selenium==4.9.1
webdriver-manager==3.8.6
urllib3==1.26.6
pytest-xdist==3.3.1
pandas==2.0.2
openpyxl==3.1.2
pytest-rerunfailures==11.1.2
jsonschema==4.17.3
pytest_unordered==0.5.2